% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{AddTreeAttribute}
\alias{AddTreeAttribute}
\title{Add attribute into the tree}
\usage{
AddTreeAttribute(d, fac, leafContent)
}
\arguments{
\item{d}{dendrogram obj}

\item{fac}{factor contains cell information and its correspondent species}

\item{leafContent}{cells in leaf nodes}
}
\value{
dendrogram with related attributes added into the tree
}
\description{
Add attribute into the tree
}
