% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{AllUpperLevelInfo}
\alias{AllUpperLevelInfo}
\title{Get all upperlevel annotations}
\usage{
AllUpperLevelInfo(d, cellannot, leafContent, propCutoff = 0.15)
}
\description{
Get all upperlevel annotations
}
