% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{assign_values_to_nodes}
\alias{assign_values_to_nodes}
\title{assign attribute value to dendrogram}
\usage{
assign_values_to_nodes(dend, attribute, value)
}
\arguments{
\item{dend}{dendrogram obj}

\item{attribute}{attribute name}

\item{value}{vector contains attribute values}
}
\value{
dendrogram obj with new attributes added to each node
}
\description{
assign attribute value to dendrogram
}
