% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{bestClusterTreeThresholds}
\alias{bestClusterTreeThresholds}
\title{caculate jacard coefficiency from best maching tree}
\usage{
bestClusterTreeThresholds(res, leaf.factor, clusters, clmerges = NULL)
}
\arguments{
\item{res}{either subsampled hierarchical graphs or a list contains merge matrix that have the similar format as igraph:::complete.dend}
}
\description{
caculate jacard coefficiency from best maching tree
}
