% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{buildSpeciesTree}
\alias{buildSpeciesTree}
\title{recursive prunning tree by minimal mixing}
\usage{
buildSpeciesTree(
  dend,
  subsamples = NULL,
  cls.groups,
  cellannot,
  species,
  upperlevelannot = NULL,
  renameCluster = TRUE,
  plot = TRUE
)
}
\arguments{
\item{dend}{dendrogram obj of original tree}

\item{subsamples}{list contains subsampled dendrograms and correspondent cell clusters}

\item{cls.groups}{clusters of original dendrograms}

\item{cellannot}{cell annotations}

\item{species}{factor annoted which cell belong to which species}

\item{upperlevelannot}{higher-level cell annotations}

\item{plot}{plot the built tree or not}
}
\description{
build entire tree based on dendrogram and subsampled dendrograms
}
