% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{cluster.matrix.expression.distances}
\alias{cluster.matrix.expression.distances}
\title{Evaluate expression distances between clusters}
\usage{
cluster.matrix.expression.distances(
  counts,
  groups = NULL,
  useVariablegenes = F,
  variableGenes = NULL,
  dist = "cor",
  use.single.cell.comparisons = FALSE,
  use.scaled.data = FALSE,
  min.cluster.size = 1,
  max.n.cells = Inf,
  n.cells = 200
)
}
\arguments{
\item{counts}{raw count matrix - rows are cells and columns are genes}
}
\description{
Evaluate expression distances between clusters
}
