% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{clusterMappingSummary}
\alias{clusterMappingSummary}
\title{cross-species mapping summary: 1:1, or many:many}
\usage{
clusterMappingSummary(
  leafannot,
  prefix = "bi",
  speciesNames = c("marmo", "human"),
  cutoff = 0.01
)
}
\description{
cross-species mapping summary: 1:1, or many:many
}
