% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{dendSetColorByMixing}
\alias{dendSetColorByMixing}
\title{Color tree based on mixing of species}
\usage{
dendSetColorByMixing(d, fac, leafContent, colorpallete)
}
\arguments{
\item{d}{dendrogram obj}

\item{fac}{factor contains species and barcodes information}

\item{colorpallete}{color pallete to color the tree
e.g. colorRampPalette(c("blue", "grey", "grey", "grey", "red"))(101)}
}
\value{
colored dendrogram
}
\description{
Color tree based on mixing of species
}
