% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{dendSetColorByNormMixing}
\alias{dendSetColorByNormMixing}
\title{colored tree by species mixing based on normalized value}
\usage{
dendSetColorByNormMixing(d, withinFacNorm = FALSE)
}
\arguments{
\item{d}{dendrogram}

\item{withinFacNorm}{if colored the tree based on wihinfactor normalization}
}
\value{
colored dendrogram
}
\description{
colored tree by species mixing based on normalized value
}
