% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{dendSetWidthBysize}
\alias{dendSetWidthBysize}
\title{Set width of tree base on size of clusters}
\usage{
dendSetWidthBysize(d, scale = 10)
}
\arguments{
\item{d}{dendrogram obj}

\item{scale}{scales to set branch size}
}
\value{
dendrogram with the width of branches indicates cluster size
}
\description{
Set width of tree base on size of clusters
}
