% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{getLeafcontentFlat}
\alias{getLeafcontentFlat}
\title{get leaf information from cell annotation}
\usage{
getLeafcontentFlat(cls.groups)
}
\arguments{
\item{cls.groups}{factor contains celltype annotation for each cell}
}
\description{
get leaf information from cell annotation
}
