% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{getUpperLevelNode}
\alias{getUpperLevelNode}
\title{get highest node with upperlevel information}
\usage{
getUpperLevelNode(dend, cutoff = 0.7)
}
\description{
get highest node with upperlevel information
}
