% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{mappingLabel}
\alias{mappingLabel}
\title{mapping leaf node according to cell annotation}
\usage{
mappingLabel(d, leafcontent, cellannot, humanAnnot = T)
}
\arguments{
\item{d}{dendrogram}

\item{cellannot}{factor contains celltype annotation for each cell}
}
\value{
leaf annotation and purity measurement
}
\description{
mapping leaf node according to cell annotation
}
