% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{MarkMixedNode}
\alias{MarkMixedNode}
\title{caculate branch entropy attribute for normalized value}
\usage{
MarkMixedNode(d, mixing.cutoff = 0.3)
}
\arguments{
\item{d}{dendrogram}
}
\value{
add entropy attributes into the dendrogram
}
\description{
caculate branch entropy attribute for normalized value
}
