% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{NormTree}
\alias{NormTree}
\title{normalize to upperlevel}
\usage{
NormTree(d, upperLevelnodes, cellannot, fac, cutoff = 0.7)
}
\arguments{
\item{d:}{dendrogram}

\item{cellannot:}{factor contains upperlevel annotation for each cell}

\item{leafContent:}{leafcontent structure for each cluster}
}
\value{
dendrogram with normalized attributes
}
\description{
normalize to upperlevel
}
