% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{NormTreeWithinFactor}
\alias{NormTreeWithinFactor}
\title{Normalize tree based on within-factors}
\usage{
NormTreeWithinFactor(d, fac, facprefix = c("human", "marmoset", "mouse"))
}
\arguments{
\item{fac}{factor contain all cells and their correspondent annotation}

\item{dend}{dendrogram obj}
}
\description{
Normalize tree based on within-factors
}
