% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{pruneTree}
\alias{pruneTree}
\title{prune tree based on mixing}
\usage{
pruneTree(dend, minCutoff = 0.3, maxCutoff = 0.4)
}
\arguments{
\item{dend}{dendrogram obj}

\item{minCutoff}{minimal value for mixing cutoff}

\item{maxCutoff}{max value for mixing cutoff}
}
\description{
prune tree based on mixing
}
