% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{pruneTreeEntropy}
\alias{pruneTreeEntropy}
\title{prune tree based on entropy}
\usage{
pruneTreeEntropy(dend, cutoff = 2.9)
}
\arguments{
\item{dend}{dendrogram obj}

\item{cutoff}{cutoff value}
}
\description{
prune tree based on entropy
}
