% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{rleiden.detection}
\alias{rleiden.detection}
\title{Run recursive leiden clustering algorithm - modified from conos (https://github.com/hms-dbmi/conos)}
\usage{
rleiden.detection(
  graph,
  K = 2,
  renameCluter = TRUE,
  n.cores = 1,
  min.community.size = 10,
  verbose = FALSE,
  resolution = 1,
  K.current = 1,
  ...
)
}
\arguments{
\item{graph}{igraph object}

\item{K}{number of iterations}

\item{renameCluster}{if rename clusters}
}
\description{
Run recursive leiden clustering algorithm - modified from conos (https://github.com/hms-dbmi/conos)
}
