% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{subSampleTree}
\alias{subSampleTree}
\title{generate subsampled dendrograms from gene expression matrix}
\usage{
subSampleTree(
  counts,
  cls.groups = NULL,
  subsample.groups = NULL,
  stability.subsamples = 10,
  Variablegenes = NULL,
  stability.subsampling.fraction = 0.95,
  renameCluster = FALSE
)
}
\arguments{
\item{counts}{scaled count matrix - rows are cells, colums are genes}

\item{cls.group}{original group to subsample}

\item{subsample.group}{list contains subsampled groups}
}
\description{
generate subsampled dendrograms from gene expression matrix
}
