% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{TreeEntropy}
\alias{TreeEntropy}
\title{caculate branch entropy attribute for normalized value}
\usage{
TreeEntropy(d, entropy.cutoff = 2, withnFacNorm = FALSE)
}
\arguments{
\item{d}{dendrogram}
}
\value{
add entropy attributes into the dendrogram
}
\description{
caculate branch entropy attribute for normalized value
}
