% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{TreeStability}
\alias{TreeStability}
\title{return stability score based on walktrap in hierachical structure}
\usage{
TreeStability(g, dend, cls.groups, cls.subsamples,
  stability.subsamples = 10, stability.subsampling.fraction = 0.95,
  min.group.size = 30, n.cores = 10)
}
\description{
return stability score based on walktrap in hierachical structure
}
