% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{TreeStabilityDend}
\alias{TreeStabilityDend}
\title{caculate stability only based on dendrogram and subsampled dendrograms}
\usage{
TreeStabilityDend(
  dend,
  cls.groups,
  subsamples,
  assignValuestoNode = TRUE,
  n.cores = 5
)
}
\arguments{
\item{dend}{original dendrogram}

\item{cls.groups}{original cell clusters}

\item{subsamples}{list contains subsampled dendrograms and correspondent cell clusters}
}
\description{
caculate stability only based on dendrogram and subsampled dendrograms
}
