% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{UpperLevelInfo}
\alias{UpperLevelInfo}
\title{add upperlevel attributes into the tree}
\usage{
UpperLevelInfo(d, cellannot, leafContent, propCutoff = 0.15)
}
\description{
add upperlevel attributes into the tree
}
