% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{dendSetColor2factorNormMixing}
\alias{dendSetColor2factorNormMixing}
\title{colored tree by species mixing based on normalized value - 2-species only}
\usage{
dendSetColor2factorNormMixing(d, colorpallete)
}
\arguments{
\item{d}{dendrogram obj}

\item{colorpallete}{vector contains colors}
}
\description{
colored tree by species mixing based on normalized value - 2-species only
}
