% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{dendSetColorByNormMixing}
\alias{dendSetColorByNormMixing}
\title{colored tree by species mixing based on normalized value}
\usage{
dendSetColorByNormMixing(d, colorpallete)
}
\description{
colored tree by species mixing based on normalized value
}
