% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTree.R
\name{TreeStability}
\alias{TreeStability}
\title{return stability scores for each node by searching for optimal matching tree}
\usage{
TreeStability(
  dataobj,
  dend,
  algorithm = "expression",
  cls.groups,
  cls.subsamples,
  stability.subsamples = 10,
  stability.subsampling.fraction = 0.95,
  min.group.size = 30,
  n.cores = 10
)
}
\arguments{
\item{dataobj}{igraph object if algorithm=walktrap, conos or gene expression matrix if algorithm=expression}

\item{dend}{original dendrogram}

\item{algorithm}{algorithms for contruct hierachical tree in subsamples: walktrap or based on expression (expression)}
}
\description{
return stability scores for each node by searching for optimal matching tree
}
